virt-api: Extract volume handlers in their own file
Renamed `getDryRunOption` to `withDryRun` in
subresource_test.go avoiding name conflicts
